sidecar - refactor tracing logic in interface.rs to separate files
