+++ sidecar/src/service/sidecar_server.rs
@@ -2,15 +2,15 @@
 // SPDX-License-Identifier: Apache-2.0
 
 use crate::config::get_product_endpoint;
-use crate::interface::{SidecarStats, TraceFlusher};
+use crate::interface::SidecarStats;
 use crate::log;
 use crate::log::{MULTI_LOG_FILTER, MULTI_LOG_WRITER};
 use crate::service::EnqueuedTelemetryData;
 use crate::service::{
-    sidecar_interface::ServeSidecarInterface, telemetry::AppInstance, AppOrQueue, InstanceId,
-    QueueId, RequestIdentification, RequestIdentifier, RuntimeInfo, RuntimeMetadata,
-    SerializedTracerHeaderTags, SessionConfig, SessionInfo, SidecarAction, SidecarInterface,
-    SidecarInterfaceRequest, SidecarInterfaceResponse,
+    sidecar_interface::ServeSidecarInterface, telemetry::AppInstance, tracing::TraceFlusher,
+    AppOrQueue, InstanceId, QueueId, RequestIdentification, RequestIdentifier, RuntimeInfo,
+    RuntimeMetadata, SerializedTracerHeaderTags, SessionConfig, SessionInfo, SidecarAction,
+    SidecarInterface, SidecarInterfaceRequest, SidecarInterfaceResponse,
 };
 use datadog_ipc::platform::{AsyncChannel, ShmHandle};
 use datadog_ipc::tarpc;
