+++ sidecar/src/service/tracing/mod.rs
@@ -1,0 +1,8 @@
+// Copyright 2021-Present Datadog, Inc. https://www.datadoghq.com/
+// SPDX-License-Identifier: Apache-2.0
+
+pub(crate) use trace_flusher::TraceFlusher;
+
+use trace_send_data::TraceSendData;
+pub(crate) mod trace_flusher;
+pub(crate) mod trace_send_data;
